Refactor PR review dialog
