+++ app/styles/_ui.scss
@@ -95,7 +95,6 @@
 @import 'ui/check-runs/_ci-check-run-popover';
 @import 'ui/check-runs/ci-check-run-job-steps';
 @import 'ui/_pull-request-checks-failed';
-@import 'ui/_pull-request-review';
 @import 'ui/_sandboxed-markdown';
 @import 'ui/_pull-request-quick-view';
 @import 'ui/discard-changes-retry';
